{{ cookiecutter.repo_name }}/requirements/base.txt
boto==2.38.0
django-bower==5.0.4
django-braces==1.8.0
django-celery==3.1.16
django-filter==0.9.2
django-formtools==1.0
django-pipeline==1.5.0
django-robots==1.0
django-storages==1.1.8
djangorestframework==3.1.1
djrill==1.3.0
flower==0.8.2
librabbitmq==1.6.1
pep8==1.6.2
Pillow==2.8.1
python-decouple==2.3
python-memcached==1.54
pytz==2015.2
raven==5.2.0
Unipath==1.1
